do not make system.histogram_metrics use Float64 for value

+++ src/Common/Histogram.h
@@ -11,7 +11,7 @@
 
 namespace DB::Histogram
 {
-    using Value = Float64;
+    using Value = Int64;
     using Buckets = std::vector<Value>;
     using Labels = std::vector<String>;
     using LabelValues = std::vector<String>;
